pyproject.toml
[project]
name = "detectormlops"
version = "0.1.0"
description = ""
authors = [
    {name = "Levanov Valeriy"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc (>=3.64.2,<4.0.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "datasets (>=4.4.1,<5.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "wandb (>=0.23.1,<0.24.0)",
    "torch (>=2.9.1,<3.0.0)",
    "accelerate (>=1.12.0,<2.0.0)",
    "lightning (>=2.6.0,<3.0.0)",
    "onnxscript (>=0.5.6,<0.6.0)",
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "tritonclient[http] (>=2.63.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "dvc-ssh (>=4.2.2,<5.0.0)"
]

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"
[tool.poetry]
packages = [{ include = "modules", from = "src" }]


[tool.poetry.dependencies]
torch = {source = "pytorch-cpu"}
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
